// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AssetType = {
  SNIPPET: 'SNIPPET',
  PROMPT: 'PROMPT',
  NOTE: 'NOTE',
  LINK: 'LINK',
  IMAGE: 'IMAGE',
  FILE: 'FILE'
} as const

export type AssetType = (typeof AssetType)[keyof typeof AssetType]